TST/REF: collect Index tests by method (#39873)
