+++ pandas/tests/indexes/common.py
@@ -243,11 +243,6 @@ class Base:
         with pytest.raises(TypeError, match=msg):
             index.copy(name=[["mario"]])
 
-    def test_copy_dtype_deprecated(self, index):
-        # GH35853
-        with tm.assert_produces_warning(FutureWarning, check_stacklevel=False):
-            index.copy(dtype=object)
-
     def test_ensure_copied_data(self, index):
         # Check the "copy" argument of each Index.__new__ is honoured
         # GH12309
@@ -504,10 +499,9 @@ class Base:
         assert empty_idx.format() == []
         assert empty_idx.format(name=True) == [""]
 
-    def test_hasnans_isnans(self, index):
+    def test_hasnans_isnans(self, index_flat):
         # GH 11343, added tests for hasnans / isnans
-        if isinstance(index, MultiIndex):
-            return
+        index = index_flat
 
         # cases in indices doesn't include NaN
         idx = index.copy(deep=True)
