+++ pandas/tests/indexes/datetimes/methods/test_insert.py
