+++ pandas/tests/indexes/datetimes/test_datetime.py
@@ -240,10 +240,3 @@ class TestDatetimeIndex:
         result = np.asarray(idx, dtype=object)
 
         tm.assert_numpy_array_equal(result, expected)
-
-    def test_to_frame_datetime_tz(self):
-        # GH 25809
-        idx = date_range(start="2019-01-01", end="2019-01-30", freq="D", tz="UTC")
-        result = idx.to_frame()
-        expected = DataFrame(idx, index=idx)
-        tm.assert_frame_equal(result, expected)
