+++ pandas/tests/indexes/period/test_period.py
@@ -79,25 +79,6 @@ class TestPeriodIndex(DatetimeLike):
         series = Series(1, index=index)
         assert isinstance(series, Series)
 
-    def test_shallow_copy_empty(self):
-        # GH13067
-        idx = PeriodIndex([], freq="M")
-        result = idx._view()
-        expected = idx
-
-        tm.assert_index_equal(result, expected)
-
-    def test_shallow_copy_disallow_i8(self):
-        # GH-24391
-        pi = period_range("2018-01-01", periods=3, freq="2D")
-        with pytest.raises(AssertionError, match="ndarray"):
-            pi._shallow_copy(pi.asi8)
-
-    def test_shallow_copy_requires_disallow_period_index(self):
-        pi = period_range("2018-01-01", periods=3, freq="2D")
-        with pytest.raises(AssertionError, match="PeriodIndex"):
-            pi._shallow_copy(pi)
-
     def test_view_asi8(self):
         idx = PeriodIndex([], freq="M")
 
@@ -411,7 +392,7 @@ class TestPeriodIndex(DatetimeLike):
         result = Index(periods)
         assert isinstance(result, PeriodIndex)
 
-    def test_append_concat(self):
+    def test_append_concat(self):  # TODO: pd.concat test
         # #1815
         d1 = date_range("12/31/1990", "12/31/1999", freq="A-DEC")
         d2 = date_range("12/31/2000", "12/31/2009", freq="A-DEC")
@@ -441,13 +422,6 @@ class TestPeriodIndex(DatetimeLike):
         result = index.map(lambda x: x.ordinal)
         exp = Index([x.ordinal for x in index])
         tm.assert_index_equal(result, exp)
-
-    def test_insert(self):
-        # GH 18295 (test missing)
-        expected = PeriodIndex(["2017Q1", NaT, "2017Q2", "2017Q3", "2017Q4"], freq="Q")
-        for na in (np.nan, NaT, None):
-            result = period_range("2017Q1", periods=4, freq="Q").insert(1, na)
-            tm.assert_index_equal(result, expected)
 
     @pytest.mark.parametrize(
         "msg, key",
