+++ pandas/tests/indexes/test_base.py
@@ -5,7 +5,6 @@ from datetime import (
 )
 from io import StringIO
 import math
-import operator
 import re
 
 import numpy as np
@@ -573,25 +572,19 @@ class TestIndex(Base):
         with pytest.raises(TypeError, match=msg):
             right.asof(left)
 
-    def test_asof_datetime_partial(self):
-        index = date_range("2010-01-01", periods=2, freq="m")
-        expected = Timestamp("2010-02-28")
-        result = index.asof("2010-02")
-        assert result == expected
-        assert not isinstance(result, Index)
-
-    def test_nanosecond_index_access(self):
-        s = Series([Timestamp("20130101")]).values.view("i8")[0]
+    # TODO: this tests Series.asof
+    def test_asof_nanosecond_index_access(self):
+        s = Timestamp("20130101").value
         r = DatetimeIndex([s + 50 + i for i in range(100)])
-        x = Series(np.random.randn(100), index=r)
+        ser = Series(np.random.randn(100), index=r)
 
-        first_value = x.asof(x.index[0])
+        first_value = ser.asof(ser.index[0])
 
         # this does not yet work, as parsing strings is done via dateutil
         # assert first_value == x['2013-01-01 00:00:00.000000050+0000']
 
         expected_ts = np_datetime64_compat("2013-01-01 00:00:00.000000050+0000", "ns")
-        assert first_value == x[Timestamp(expected_ts)]
+        assert first_value == ser[Timestamp(expected_ts)]
 
     @pytest.mark.parametrize("index", ["string"], indirect=True)
     def test_booleanindex(self, index):
@@ -634,110 +627,6 @@ class TestIndex(Base):
         msg = r"arrays used as indices must be of integer \(or boolean\) type"
         with pytest.raises(IndexError, match=msg):
             index[empty_farr]
-
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    def test_intersection(self, index, sort):
-        first = index[:20]
-        second = index[:10]
-        intersect = first.intersection(second, sort=sort)
-        if sort is None:
-            tm.assert_index_equal(intersect, second.sort_values())
-        assert tm.equalContents(intersect, second)
-
-        # Corner cases
-        inter = first.intersection(first, sort=sort)
-        assert inter is first
-
-    @pytest.mark.parametrize(
-        "index2,keeps_name",
-        [
-            (Index([3, 4, 5, 6, 7], name="index"), True),  # preserve same name
-            (Index([3, 4, 5, 6, 7], name="other"), False),  # drop diff names
-            (Index([3, 4, 5, 6, 7]), False),
-        ],
-    )
-    def test_intersection_name_preservation(self, index2, keeps_name, sort):
-        index1 = Index([1, 2, 3, 4, 5], name="index")
-        expected = Index([3, 4, 5])
-        result = index1.intersection(index2, sort)
-
-        if keeps_name:
-            expected.name = "index"
-
-        assert result.name == expected.name
-        tm.assert_index_equal(result, expected)
-
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    @pytest.mark.parametrize(
-        "first_name,second_name,expected_name",
-        [("A", "A", "A"), ("A", "B", None), (None, "B", None)],
-    )
-    def test_intersection_name_preservation2(
-        self, index, first_name, second_name, expected_name, sort
-    ):
-        first = index[5:20]
-        second = index[:10]
-        first.name = first_name
-        second.name = second_name
-        intersect = first.intersection(second, sort=sort)
-        assert intersect.name == expected_name
-
-    def test_chained_union(self, sort):
-        # Chained unions handles names correctly
-        i1 = Index([1, 2], name="i1")
-        i2 = Index([5, 6], name="i2")
-        i3 = Index([3, 4], name="i3")
-        union = i1.union(i2.union(i3, sort=sort), sort=sort)
-        expected = i1.union(i2, sort=sort).union(i3, sort=sort)
-        tm.assert_index_equal(union, expected)
-
-        j1 = Index([1, 2], name="j1")
-        j2 = Index([], name="j2")
-        j3 = Index([], name="j3")
-        union = j1.union(j2.union(j3, sort=sort), sort=sort)
-        expected = j1.union(j2, sort=sort).union(j3, sort=sort)
-        tm.assert_index_equal(union, expected)
-
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    def test_union(self, index, sort):
-        first = index[5:20]
-        second = index[:10]
-        everything = index[:20]
-
-        union = first.union(second, sort=sort)
-        if sort is None:
-            tm.assert_index_equal(union, everything.sort_values())
-        assert tm.equalContents(union, everything)
-
-    @pytest.mark.parametrize("klass", [np.array, Series, list])
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    def test_union_from_iterables(self, index, klass, sort):
-        # GH 10149
-        first = index[5:20]
-        second = index[:10]
-        everything = index[:20]
-
-        case = klass(second.values)
-        result = first.union(case, sort=sort)
-        if sort is None:
-            tm.assert_index_equal(result, everything.sort_values())
-        assert tm.equalContents(result, everything)
-
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    def test_union_identity(self, index, sort):
-        first = index[5:20]
-
-        union = first.union(first, sort=sort)
-        # i.e. identity is not preserved when sort is True
-        assert (union is first) is (not sort)
-
-        # This should no longer be the same object, since [] is not consistent,
-        # both objects will be recast to dtype('O')
-        union = first.union([], sort=sort)
-        assert (union is first) is (not sort)
-
-        union = Index([]).union(first, sort=sort)
-        assert (union is first) is (not sort)
 
     def test_union_dt_as_obj(self, sort):
         # TODO: Replace with fixturesult
@@ -865,123 +754,6 @@ class TestIndex(Base):
         result = left.append(right)
         assert result.name == expected
 
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    @pytest.mark.parametrize("second_name,expected", [(None, None), ("name", "name")])
-    def test_difference_name_preservation(self, index, second_name, expected, sort):
-        first = index[5:20]
-        second = index[:10]
-        answer = index[10:20]
-
-        first.name = "name"
-        second.name = second_name
-        result = first.difference(second, sort=sort)
-
-        assert tm.equalContents(result, answer)
-
-        if expected is None:
-            assert result.name is None
-        else:
-            assert result.name == expected
-
-    def test_difference_empty_arg(self, index, sort):
-        first = index[5:20]
-        first.name = "name"
-        result = first.difference([], sort)
-
-        tm.assert_index_equal(result, first)
-
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    def test_difference_identity(self, index, sort):
-        first = index[5:20]
-        first.name = "name"
-        result = first.difference(first, sort)
-
-        assert len(result) == 0
-        assert result.name == first.name
-
-    @pytest.mark.parametrize("index", ["string"], indirect=True)
-    def test_difference_sort(self, index, sort):
-        first = index[5:20]
-        second = index[:10]
-
-        result = first.difference(second, sort)
-        expected = index[10:20]
-
-        if sort is None:
-            expected = expected.sort_values()
-
-        tm.assert_index_equal(result, expected)
-
-    @pytest.mark.parametrize("opname", ["difference", "symmetric_difference"])
-    def test_difference_incomparable(self, opname):
-        a = Index([3, Timestamp("2000"), 1])
-        b = Index([2, Timestamp("1999"), 1])
-        op = operator.methodcaller(opname, b)
-
-        with tm.assert_produces_warning(RuntimeWarning):
-            # sort=None, the default
-            result = op(a)
-        expected = Index([3, Timestamp("2000"), 2, Timestamp("1999")])
-        if opname == "difference":
-            expected = expected[:2]
-        tm.assert_index_equal(result, expected)
-
-        # sort=False
-        op = operator.methodcaller(opname, b, sort=False)
-        result = op(a)
-        tm.assert_index_equal(result, expected)
-
-    @pytest.mark.xfail(reason="Not implemented")
-    @pytest.mark.parametrize("opname", ["difference", "symmetric_difference"])
-    def test_difference_incomparable_true(self, opname):
-        # TODO decide on True behaviour
-        # # sort=True, raises
-        a = Index([3, Timestamp("2000"), 1])
-        b = Index([2, Timestamp("1999"), 1])
-        op = operator.methodcaller(opname, b, sort=True)
-
-        with pytest.raises(TypeError, match="Cannot compare"):
-            op(a)
-
-    def test_symmetric_difference_mi(self, sort):
-        index1 = MultiIndex.from_tuples(zip(["foo", "bar", "baz"], [1, 2, 3]))
-        index2 = MultiIndex.from_tuples([("foo", 1), ("bar", 3)])
-        result = index1.symmetric_difference(index2, sort=sort)
-        expected = MultiIndex.from_tuples([("bar", 2), ("baz", 3), ("bar", 3)])
-        if sort is None:
-            expected = expected.sort_values()
-        tm.assert_index_equal(result, expected)
-        assert tm.equalContents(result, expected)
-
-    @pytest.mark.parametrize(
-        "index2,expected",
-        [
-            (Index([0, 1, np.nan]), Index([2.0, 3.0, 0.0])),
-            (Index([0, 1]), Index([np.nan, 2.0, 3.0, 0.0])),
-        ],
-    )
-    def test_symmetric_difference_missing(self, index2, expected, sort):
-        # GH 13514 change: {nan} - {nan} == {}
-        # (GH 6444, sorting of nans, is no longer an issue)
-        index1 = Index([1, np.nan, 2, 3])
-
-        result = index1.symmetric_difference(index2, sort=sort)
-        if sort is None:
-            expected = expected.sort_values()
-        tm.assert_index_equal(result, expected)
-
-    def test_symmetric_difference_non_index(self, sort):
-        index1 = Index([1, 2, 3, 4], name="index1")
-        index2 = np.array([2, 3, 4, 5])
-        expected = Index([1, 5])
-        result = index1.symmetric_difference(index2, sort=sort)
-        assert tm.equalContents(result, expected)
-        assert result.name == "index1"
-
-        result = index1.symmetric_difference(index2, result_name="new_name", sort=sort)
-        assert tm.equalContents(result, expected)
-        assert result.name == "new_name"
-
     def test_is_mixed_deprecated(self):
         # GH#32922
         index = self.create_index()
@@ -1035,7 +807,7 @@ class TestIndex(Base):
             assert index.is_all_dates is expected
 
     def test_summary(self, index):
-        self._check_method_works(Index._summary, index)
+        index._summary()
 
     def test_summary_bug(self):
         # GH3869`
@@ -1097,9 +869,6 @@ class TestIndex(Base):
     def test_logical_compat(self, op):
         index = self.create_index()
         assert getattr(index, op)() == getattr(index.values, op)()
-
-    def _check_method_works(self, method, index):
-        method(index)
 
     @pytest.mark.parametrize("index", ["string", "int", "float"], indirect=True)
     def test_drop_by_str_label(self, index):
@@ -1776,11 +1545,6 @@ class TestMixedIntIndex(Base):
         with pytest.raises(ValueError, match=msg):
             Index([1, 2, 3]).dropna(how="xxx")
 
-    def test_get_combined_index(self):
-        result = _get_combined_index([])
-        expected = Index([])
-        tm.assert_index_equal(result, expected)
-
     @pytest.mark.parametrize(
         "index",
         [
@@ -1812,16 +1576,6 @@ class TestMixedIntIndex(Base):
         msg = "^'str' object cannot be interpreted as an integer$"
         with pytest.raises(TypeError, match=msg):
             bytes(index)
-
-    def test_intersect_str_dates(self):
-        dt_dates = [datetime(2012, 2, 9), datetime(2012, 2, 22)]
-
-        index1 = Index(dt_dates, dtype=object)
-        index2 = Index(["aa"], dtype=object)
-        result = index2.intersection(index1)
-
-        expected = Index([], dtype=object)
-        tm.assert_index_equal(result, expected)
 
     @pytest.mark.filterwarnings("ignore:elementwise comparison failed:FutureWarning")
     def test_index_with_tuple_bool(self):
@@ -1866,6 +1620,11 @@ class TestIndexUtils:
         ]
         result = ensure_index(intervals)
         expected = Index(intervals, dtype=object)
+        tm.assert_index_equal(result, expected)
+
+    def test_get_combined_index(self):
+        result = _get_combined_index([])
+        expected = Index([])
         tm.assert_index_equal(result, expected)
